Unset validation-disabling property.
To see the current classpath do mvn dependency:list -Preleases. We pull in javax.validation as Spring Data Neo4j has a dependency on it. That causes Hibernate to try to bootstrap the javax.validation integration and fails as it cannot lookup a default provider.

To reproduce run mvn clean test -Preleases. Current master has the property enabled and runs fine.